requirements.txt
beautifulsoup4>=4.12,<5.0
faiss-cpu>=1.7,<2.0
nltk>=3.8,<4.0
python-docx>=1.1,<2.0
scikit-learn>=1.4,<2.0
scipy>=1.11,<2.0
